backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
httpx==0.25.2
google-generativeai==0.8.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
alembic==1.13.0
aiofiles==23.2.1
python-dateutil==2.8.2
requests==2.31.0
pandas==2.1.4
pydantic-settings==2.0.3 